NSFW topic?
Added a topic about when the player types in lewd stuff, f*pp*ng to be specific.
Since this is an unavoidable topic, I thought it would be alright to have Monika have a response towards the thought, but still keeping it PG.
In fact, she urges to player not to, since it had bad side-effects plus the fact that she doesn't want the player to think of other girls besides her.

Removed from the random topics, since this would be an awkward topic to talk about with Monika. So... only the people that would type the keywords can read this dialogue.

Please give this topic a complete review, if this is an appropriate topic to add to Monika's roster. I'm also okay if this would be edited to further suit Monika's character, IF this is not appropriate. Thank you.

Topic starts from line 2720 and ends at line 2753.