Greeting for Player.Chr?
The actual dialogue in this greeting is more than up for debate, but I spent some time on working on a few lines of code that could detect if someone added a .chr file with the same name as their playername. Could be a good transition into 'Monika makes her own route', or just make room for all sorts of different easter eggs.