Classic-like pong mechanics
Makes the pong game feel more like the classic one. The ball reflects
off the paddles at an angle that depends on the point of impact on
the paddle. Hitting dead-centre makes the ball go horizontally, while
hitting the corners makes it go at a large angle (currently 60 deg).
Also changes the ball to not make bounce sounds when it's still stuck
on the paddle.