editing music topics
modified the instrument question in monika_orchestra to accomodate for if the player plays piano. (possible issue with input: what about one-word responses with punctuation on the end? or longer responses?)

also, re.: having the instrument question only be asked on the first encounter with the topic: is persistent.instrument set to anything prior to the first answering of the question? if it's something like None, then there could be an if statement along the lines of "if persistent.instrument == None:" and so on. there might be something i'm overlooking, but thought i'd say it anyhow.

also added monika_jazz