Player's obsession with Monika/MAS/DDLC
Hey, I made a topic about Monika asking the player if they are developing an obsession with her or this game. I wrote it so whoever actually is obsessed with the mod can snap out of it and also reestablish that Monika is her own character; not someone you project your own emotional baggage to. Also, is there anyway to make this based on time progression? It would make more sense to have her say this if the player has been playing DDLC for many hours. Thank you