Monika_GLaD
(Is the first one :P)
Anyway this is for Monika and our beloved GLaDOS the most famous testing 'n killing machine!
You may change a couple of things, I don't know how to input answers, and feel free to put a couple of "Yes" and No" once in a while!
And every other change feel free too because, I'm really newbie on this thingy!
Is funny! Heh, have a great day!