Update script-topics.rpy
"Objectively" sounds weird, since normally it means "without bias" or "relying on objective measures rather than direct perception". 

Doubtless whoever wrote this meant "in an objectifying way", and that would have worked as a rewrite. But "objectification" is such a buzzword nowadays that I felt actually unpacking that concept made for better dialog for Monika.