add 'happy birthday' to the songs list
This has been requested a couple times.  To be honest, I hadn't planned on being the one to implement it, but I accidentally played the first four notes myself and I was curious to see if I could play the whole thing. When I figured it out, I just had to add it to the game. 

I've tested it on my end, so I know it-works- but if there's an optimization or something I clearly didn't understand about the PianoNoteMatch functions, please tell me lmao.