3 Additional Moods
The 2 others are small yet nice little additions to portraying a players feelings. With the Guilty mood, it's proposed as shown by the dialogue, that the player confesses a troubling thought to Monika. Depending on how long or how much they say (based on time frame), will they get alternate responses.