2 Topic Additions
The first on pokes fun at how love at first sight would practically work, but also a reminder to how Monika came to loving the player in the first place. The other is a discussion on the artstyle that goes into anime, how unique and beautiful it can look, but also points out how unrealistic it is, and refers back to how Monika was drawn the same way.