New Graduation Topic/Speech
Ok, first things first. MASSIVE thanks to @jmwall24 for: 
1. Proposing the idea
2. Helping with expressions
3. Helping edit the speech
4. Helping with testing
5. Lots of ideas
6. Loads of dialogue writing too

Anyway. We figured that Monika *would* have been the one to graduate at the top of her class, so she'd be giving a speech for the graduating class, so we wrote one, complete with music and probably everything else you can think of. That said, since this also requires music, here's a file (https://drive.google.com/open?id=1g2q3UqxfkG3jGXY1RvR6KIFETThJLmSb) which will need to go in `/mod_assets/sounds/amb` for this to actually play.