Topic on time travel and changing the system date
I know that people have tried changing the system date in order to unlock certain features in MAS that are only available on a specific day or after a certain period of time. I also know that this kind of "time travel" has led to these people having problems. So I wrote a topic that segues nicely from a discussion about the effects of time travel on the people closest to the time traveller to a warning about the effects of changing the system date.

I am open to suggestions for how to improve this topic. In particular, if anyone knows any books whose main focus is how people are affected when someone close to them goes time travelling, I'd be glad to have an example to mention in this topic.