Fixed? Maybe?
I think I did it right. 
referring to other scripts in another script is something I'm relativity new to. So sorry if it takes a minute for me to fully understand.
If it's right, awesome. If not, I'll fix it. 

Heading to work here in a minute, but I'll be able to finish when I get home.