4chan Window reaction
I feel obligated to add this to help close the "pewdiepipeline" and prevent terror wherever nessicary. If you do not know what that is, I encourage you to watch this:https://youtu.be/pnmRYRRDbuw I hope you consider this addition or a version of it that conveys this message better as I'm sure that monika would. Thank you.

P.S. I hope this is the right way to send a Pull request.