Yandere Or Not
Added a new variable: persistent._mas_pm_likes_yanderes, signed integer, that is set to 1, if player is into yanderes, 0 if the player is neutral and -1, if the player is scared of them. Default is 0. Now during the "Yandere Yuri" topic Monika will ask, if the player would like her to be more yandere-ish... then it will set the persistent._mas_pm_likes_yanderes to 1, if the player responds "Actually... I'm into yanderes", leave it at 0, if the player responds "No" or set it to -1, if the player says "HELL NO". "Yandere Yuri" topic now returns derandom. Depending on the persistent._mas_pm_likes_yanderes value the following conversations will be affected:

- monika_simulated
- monika_girlfriend
- monika_waifus
- monika_herself
- monika_harem
- monika_hack
- monika_swordsmanship

I will also change the script about watching player while he's asleep in script-moods.rpy