I Look Up To You - New Compliment
Hi!! This is my first attempt at contributing to the mod through code (that wasn't a .json file!). I'm sorry if this has some issues, and I have a feeling someone else had already done this... But it was still fun! If I need to add a menu, let me know!