"Sweater Curse" topic
I'm pretty sure I made the PR correctly this time, here goes nothing.
This one is a silly little thing for winter, an odd little superstition that actually has some possible psychological roots.

I was originally going to make this one a winter-only random topic, but I wasn't 100% sure so I left it as normal for now. I'm happy to add that in if folks think that's a better idea!
I'm also considering having it derandom after it's viewed once, since it's a bit long? I'd appreciate opinions!