I feel... Like Crying.
Please don’t murder me with how short this is - I’m not done yet, and it’s 2:30 in the morning and I need sleep :”D

But while I still had my computer I had programmed this and a bunch of other moods into the game. This one in particular was very special to me, and I’d thought I’d put it out here for other people as well.

If this doesn’t fit, that’s okay. Maybe I can make it a submod or something. It serves a similar purpose to sad, but it just felt more therapeutic to me somehow.