Rewrote the last part
I hope I made some changed you guys liked, but of course, there's no rush about this, feel free to comment on every issue you have with the new version, and propose your own changes too!

The idea behind this change was to focus less on looks, and more about the entire character and portrayal of a person, while also removing the whole story behind every scar narrative, that was a bit on the nose for some of us.