Rewriting of the entire text
This change was suggested by Multimokia, and was aimed at using the advice left by the other contributors, that was supposed to encompass a more general audience so no one's really left out, and maybe it'll help reassure everyone who reads it.

Spritecodes will be revised soon, as all of them may not exist yet.