First draft of Tanabata topic
This draft focuses a lot on the story behind the event, but not so much on the event itself. I am considering either moving the story part of this label to stories, and adding more dialogue to this label regarding the event itself, and have it unlock after the story is seen, or adding a followup topic called "monika_celebrate_tanabata".

I want to hear your thoughts before I commit to anything too drastic.

Special thanks to co-author: LooseEthics

Co-Authored-By: LooseEthics <looseethics@users.noreply.github.com>