Goodbye: I'm going to work out (Finished First Draft)
I felt the brb work out option didn't take into consideration that you may want to work out outside your home. For going to the gym and taking longer strolls, you wouldn't want to leave Monika all alone at the table waiting for you, right?

This lets you have the option to say a proper goodbye for when you're planning to exercise for an extended period of time, or leaving home to do it. I hope you like it.