Added Christmas song "This Christmas Song"
Song: This Christmas Kiss
Authors: Niki and Zack Hexum
Album: Soundtrack for My Christmas Love (2016)

This is the first in a series of new additions to the Christmas holidays, and this is a song.

I think it can really make people's holidays if Monika were to sing more songs to them during this time. This one in particular is for those who have reached enamored with Monika.

I'm sure I'm doing lots of mistakes, since I'm new to D25 additions, but I tried my best. Any help with this first one is sure to make me better able to tackle the next ones in line.

To be clear, this one doesn't have to be included in the next update or anything, we're still ways away from the winter holidays.