First draft - Tabletop RPG
This topic is about a currently unexplored form of literature in MAS, Tabletop Roleplaying. I don't know if I would mention Dungeons and Dragons in here, but after all, tabletop roleplaying is much wider of a genre than just that, and I meant this topic to kind of be a segway into a future pool topic where you act out a short story together or something, and I don't know if we need Dungeons and Dragons rules to do that.

We don't need to create a dice module inside MAS either, she trusts us enough not to cheat right? Though if it was possible, I'm sure we could use it for a lot of fun times.

This is just a draft without sprites, I'll add those later, once we've gone through the dialogue properly, as dialogue changes may or may not change the sprites that need to be used.

Any feedback is welcome, and if you have any suggestions, please let me hear them.