First draft: Bye: Hangout out with a friend
Any help making the dialogue better and more natural feeling is welcome, as I don't know if I really love it.

In any case, this PR is about being able to tell Monika that you're going to hang out with friends. Not sure if I should've made the prompt plural or not, but it sounds better to hang out with "friends" if you're just hanging out with one, than hanging out with "friend" if you're hanging out with multiple to me.

Please add discussion tag so we can work this out :)