Added song: I try
Song title: I try
Artist: Macy Gray
Release date: 1999

This song was part of my playlist growing up, and I always thought Macy's voice and the lyrics of this song to be something quite special. It also fits Monika pretty well, given the situation we find ourselves in with her.