Added chance for Monika to ask for goodnight kiss
With this, there's a chance that Monika will ask for a goodnight kiss when kissing is unlocked and you go to sleep at a good time. Slep slep.

If there's any way to improve the kissing sequence or the dialogue, let me know. It's been tested and it seems to run fine.