Interrupting each other
Adds small chance for special dialogue with Monika if you talk just before she's about to say something.

Anything here is still subject to tweaking, but it's been tested and seems to run smoothly. At the moment I've chosen less than 2 seconds before she starts talking, so on lower random chatter settings it's going to be really hard to do on purpose. It's meant to be a very rare occurence.

The way it is currently set up, it shouldn't ever trigger when Monika is not about to say something, such as when the random chatter setting is set to "Never". This way, you know that if you trigger this piece of dialogue, Monika was actually about to start talking when you pressed "Talk".

Huge thanks to @TarossVelin for his issue suggestion #6239 , and the devs for doing a lot of the heavy lifting behind the scenes.

Co-Authored-By: TarossVelin <43471712+TarossVelin@users.noreply.github.com>