Early draft, feedback welcome
This is an early draft of the Relationship_Quiz topic. Some of the questions may be unsuitable for minors, so by viewing the contents herein you agree you are 18+ and agree to talk about adult stuff.

I appreciate any feedback, as I want this quiz to be relevant to as many people as possible. Want your question added? Just ask.