New topic — Renewable energy
First pull request, go easy on me. 

Monika seems to care a lot about the environment, so it only feels natural to give her a topic about renewable energy! I'm pretty sure this is the first one (correct me if I'm wrong) about renewable energy, but if it isn't then I apologize. My goal here was to try and keep things neutral, but to give an overall optimistic vibe. I'm pretty sure (?) this is coded to be a random convo that will be available in the technology category later, but if it isn't just let me know. I'm a writer, not a programmer <.>