Edits in response to requested changes by Rai99
- Wink in the last line replaced with closed eyes (happy)
- Several lines rewritten or added to provide a more hollistic discussion that is far less assumptive. It's gotten a lot lengthier, but maybe it's better this way.
- Dissolve added
- Minor grammatical adjustments

This isn't the final version, just a WIP revision to show what I'm thinking of in response to the requested changes. I also need to double check the poses as well.

You know, I wonder if it would be possible to have an 'environment' category in the future. We have Monika's lines on being a vegetarian, whether she'd eat meat or not, what she thinks about urbanization, and (if this goes through) her thoughts on renewable energy. I have a very WIP "Environment" prompt as well that we could add in as well.