New topic/category — Environment
Monika is known to care about the environment (as demonstrated in a couple of her topics), but why is that? The purpose of this topic is to help shed a bit of light on the subject, and to possibly foster a small category of environment-centered topics that already exist (e.g. Veggies, cities, eating meat, and my renewable energy topic if that pull request gets accepted). 

This initial commit only has the code for the topic because I wasn't sure how to add a category into the game (not sure if it happens by default if I put 'Environment' in the category list).