First draft: monika_information_age
This draft focuses on talking about the information age, reasons for its inception, and challenges we face now because of it. If you have any input, I'm willing to alter or add anything you want. Lets make another great topic!

Expressions will be added later when the dialogue is starting to take shape.