Update script-farewells.rpy
In line 441, when 'player' has to go to class under 20 minutes after opening the game Monika will say "Aw, going already?" but her expression will look rather happy than sad. Don't know if this is made on purpose or is it a mistake but I believe that more "normal" expression would be sad so I recommend changing expression code from "1eub"to "1ekd".