expression rework
Ok, so I found the "Monika Likes and Dislikes" file again and added a bit more shade to the Shakespeare topic.  (And tried to make the entire thing seem less horoscopy?  I don't believe in horoscopes and actively hate them (and blood type personalities), so I don't see why this topic would be 'off limits' because of that character trait.)  I did a bit of reworking on timing using auto (instead of clicking through...) so it flows better (I think).
Just in case github inserts tabs instead of spaces like notepad++ I just overwrote the entire block to be safe.

Let me know if I screwed up the wink, but as far as I'm concerned I /seem/ to have nailed it.