New random topic: Elegy XI
This is my first time ever attempting to make a new random topic. Is it worth working on this? I would appreciate any critics. Besides, English is not my native language, so any corrections would be nice aswell. This is just a draft, and I want an opinion before working deeper on this one.