Add topic "War in the world"
I suggested this topic back in early February (#8665), but I think I was misunderstood. I suggested not the idea of the topic (I'm sure that it was offered before), but the topic itself. But earlier I was too stupid and did not know how to pull request.
So I leave it here if you want to add something about the war.
PS: be careful with grammar, I'm sure there is a lot of mistakes, unfortunately.