[Random Topic] Gifted Kids
Adds a random topic where Monika tells the player about gifted individuals before asking the player if they are gifted. I did local testing to make sure everything works. Completely open to suggestions on dialog changes. I'm slightly on the fence about whether or not asking if the player is gifted could be perceived as rude or insensitive in some way. Thanks in advance for any feedback.