"Do you miss the other girls?" Topic.
I always thought it'd be nice to know if she sometimes misses her club members, given that they were meant to be friends prior to MC's arrival. Here's my view on the topic. Apologies for any bugs or mistakes in this, I checked everything as thoroughly as I could. Grammatical corrections are appreciated!